fix chart, add json R/W

+++ chartdatachangedtrend.cpp
@@ -13,7 +13,7 @@ bool ChartDataChangedTrend::readJson(const QString &str)
     _data_source->buffer1_data = fromJsonArray<int>(obj["buffer1_data"].toArray());
     _data_source->buffer2_data = fromJsonArray<int>(obj["buffer2_data"].toArray());
     _data_source->buffer3_data = fromJsonArray<int>(obj["buffer3_data"].toArray());
-    _data_source->time_staps = fromJsonArray<int>(obj["time_staps"].toArray());
+    _data_source->time_staps = fromJsonArray<qint64>(obj["time_staps"].toArray());
     _data_source->interval = obj["interval"].toInt();
     return true;
 }
